Context fixes for macOS.
